apps/backend/requirements.txt
fastapi==0.115.8
pydantic==2.9.2
uvicorn[standard]==0.24.0
httpx==0.27.0
psycopg[binary]==3.2.6
redis==5.0.8
jsonschema==4.22.0
pyyaml==6.0.2
pytest==8.3.4
pytest-asyncio==0.24.0
